Add option to feedback on magnetic diagnostics (#38)
* Add option to feedback on magnetic diagnostics

This the simplest possible implementation and depends on a PR for VacuumFields.

* Fix probe and flux loop setup

This has now been tested, but may not be working properly yet

* Include inverse uncertainty in weights

Also some fixes to the optional arguments and appending the boundary
cps for testing purposes.

* More careful checking of sensor validity, weight, and uncertainty

Bad sensors are not used for control points so the constraints
should better represent the experiment now and no longer have NaN/Inf
weights.

Unfortunately some more finesse will be required for the data or
weights because the results seem poor at the moment.

* Bump compat for VacuumFields

* Fix missing iso offset

* Cleanup code duplication

* Add option for non-differential flux loop fitting

Also normalizing the probe and flux loop uncertainty factors to 1

Suggestions from code review have been implemented as well.

* Generalize type for flux_loop_weights and magnetic_probe_weights

---------

Co-authored-by: Brendan Lyons <lyonsbc@fusion.gat.com>